Move the internal_model generators into a better namespace
