+++ catalogue_pipeline/merger/src/test/scala/uk/ac/wellcome/platform/merger/rules/singlepagemiro/SierraMiroPartitionerTest.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.platform.merger.rules.singlepagemiro
 
 import org.scalatest.{FunSpec, Matchers}
-import uk.ac.wellcome.models.work.test.util.WorksGenerators
+import uk.ac.wellcome.models.work.generators.WorksGenerators
 
 class SierraMiroPartitionerTest
     extends FunSpec
