+++ catalogue_pipeline/merger/src/test/scala/uk/ac/wellcome/platform/merger/services/MergerManagerTest.scala
@@ -1,13 +1,13 @@
 package uk.ac.wellcome.platform.merger.services
 
 import org.scalatest.{FunSpec, Matchers}
+import uk.ac.wellcome.models.work.generators.WorksGenerators
 import uk.ac.wellcome.models.work.internal.{
   BaseWork,
   IdentifiableRedirect,
   UnidentifiedRedirectedWork,
   UnidentifiedWork
 }
-import uk.ac.wellcome.models.work.test.util.WorksGenerators
 
 class MergerManagerTest extends FunSpec with Matchers with WorksGenerators {
