+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v2/DisplayWorkV2Test.scala
@@ -4,8 +4,8 @@ import org.scalatest.prop.PropertyChecks
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.display.models._
 import uk.ac.wellcome.models.work.internal._
-import uk.ac.wellcome.models.work.test.util.WorksGenerators
 import org.scalacheck.ScalacheckShapeless._
+import uk.ac.wellcome.models.work.generators.WorksGenerators
 
 class DisplayWorkV2Test
     extends FunSpec
