+++ sbt_common/internal_model/src/test/scala/uk/ac/wellcome/models/transformable/sierra/test/utils/SierraGenerators.scala
@@ -9,8 +9,8 @@ import uk.ac.wellcome.models.transformable.sierra.{
   SierraItemNumber,
   SierraItemRecord
 }
-import uk.ac.wellcome.models.work.test.util.IdentifiersGenerators
 import uk.ac.wellcome.json.JsonUtil._
+import uk.ac.wellcome.models.work.generators.IdentifiersGenerators
 
 import scala.util.Random
