+++ sbt_common/internal_model/src/test/scala/uk/ac/wellcome/models/work/generators/IdentifiersGenerators.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models.work.test.util
+package uk.ac.wellcome.models.work.generators
 
 import uk.ac.wellcome.models.work.internal.{IdentifierType, SourceIdentifier}
