+++ sbt_common/internal_model/src/test/scala/uk/ac/wellcome/models/work/internal/UnidentifiedRedirectedWorkTest.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.models.work.internal
 
 import org.scalatest.{FunSpec, Matchers}
-import uk.ac.wellcome.models.work.test.util.WorksGenerators
+import uk.ac.wellcome.models.work.generators.WorksGenerators
 
 class UnidentifiedRedirectedWorkTest
     extends FunSpec
